requirements/dev.txt
-r base.txt

black==19.3b0
ipdb==0.12.2
PyHamcrest==1.9.0
pytest==5.1.2
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-flake8==1.0.4
pytest-pythonpath==0.7.3
pytest-sugar==0.9.2
requests==2.22.0
mixer==6.1.3
